Reuse summary text in command line output and in Markdown feedback.
