+++ sigridci/sigridci/reports/report.py
@@ -15,9 +15,7 @@
 import urllib.parse
 from datetime import datetime
 
-from ..objective import Objective, ObjectiveStatus
 from ..platform import Platform
-from ..publish_options import PublishOptions
 
 
 class Report:
@@ -62,22 +60,6 @@ class Report:
         customer = urllib.parse.quote_plus(options.customer.lower())
         system = urllib.parse.quote_plus(options.system.lower())
         return f"{options.sigridURL}/{customer}/{system}"
-
-    def getSummaryText(self, feedback, options):
-        status = Objective.determineStatus(feedback, options)
-        targetRating = PublishOptions.DEFAULT_TARGET if isinstance(options.targetRating, str) else options.targetRating
-        targetText = f"{targetRating:.1f} stars"
-
-        if status == ObjectiveStatus.ACHIEVED:
-            return f"✅  You wrote maintainable code and achieved your objective of {targetText}"
-        elif status == ObjectiveStatus.IMPROVED:
-            return f"↗️  You improved your code's maintainability towards your objective of {targetText}"
-        elif status == ObjectiveStatus.UNCHANGED:
-            return f"⏸️️  Your rating did not change and is still below your objective of {targetText}"
-        elif status == ObjectiveStatus.WORSENED:
-            return f"⚠️  Your code did not improve towards your objective of {targetText}"
-        else:
-            return "💭️  You did not change any files that are measured by Sigrid"
 
 
 class MarkdownRenderer:
